requirements.txt
playwright==1.49.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pytest==8.3.4
